# swagger.json
{
   "swagger": "2.0",
   "info": {
      "title": "Hotel API",
      "description": "API for hotel management",
      "version": "1.0"
   },
   "produces": [
      "application/json"
   ],
   "host": "127.0.0.1:3000",
   "basePath": "/",
   "paths": {
...
      "/reservations": {
         "get": {
            "operationId": "reservations",
            "tags": [
               "Reservations"
            ],
            "description": "Récupère la liste des réservations",
            "parameters": [],
            "responses": {
               "200": {
                  "description": "successful operation",
                  "schema": {
                     "type": "array",
                     "items": {
                        "type": "object",
                        "properties": {
                           "id_reservation": {
                              "type": "integer"
                           },
                           "check_in_date": {
                              "type": "string",
                              "format": "date-time"
                           },
                           "check_out_date": {
                              "type": "string",
                              "format": "date-time"
                           },
                           "total_cost": {
                              "type": "number",
                              "format": "float"
                           },
                           "reservation_status": {
                              "type": "string"
                           },
                           "id_room": {
                              "type": "integer"
                           },
                           "id_client": {
                              "type": "integer"
                           }
                        }
                     }
                  }
               },
               "404": {
                  "description": "Not found"
               },
               "500": {
                  "description": "Internal server error"
               }
            }
         },
         "post": {
            "operationId": "reservations_create",
            "tags": [
               "Reservations"
            ],
            "description": "Crée une réservation",
            "parameters": [
               {
                  "name": "reservation",
                  "in": "body",
                  "required": true,
                  "schema": {
                     "type": "object",
                     "properties": {
                        "check_in_date": {
                           "type": "string",
                           "format": "date-time"
                        },
                        "check_out_date": {
                           "type": "string",
                           "format": "date-time"
                        },
                        "total_cost": {
                           "type": "number",
                           "format": "float"
                        },
                        "reservation_status": {
                           "type": "string"
                        },
                        "id_room": {
                           "type": "integer"
                        },
                        "id_client": {
                           "type": "integer"
                        }
                     }
                  }
               }
            ],
            "responses": {
               "201": {
                  "description": "successful operation",
                  "schema": {
                     "type": "object",
                     "properties": {
                        "id_reservation": {
                           "type": "integer"
                        },
                        "check_in_date": {
                           "type": "string",
                           "format": "date-time"
                        },
                        "check_out_date": {
                           "type": "string",
                           "format": "date-time"
                        },
                        "total_cost": {
                           "type": "number",
                           "format": "float"
                        },
                        "reservation_status": {
                           "type": "string"
                        },
                        "id_room": {
                           "type": "integer"
                        },
                        "id_client": {
                           "type": "integer"
                        }
                     }
                  }
               }
            }
         }
      },
...
      "/services": {
         "get": {
            "operationId": "services",
            "tags": [
               "Services"
            ],
            "description": "Récupère la liste des services",
            "parameters": [],
            "responses": {
               "200": {
                  "description": "successful operation",
                  "schema": {
                     "type": "array",
                     "items": {
                        "type": "object",
                        "properties": {
                           "id_service": {
                              "type": "integer"
                           },
                           "service_name": {
                              "type": "string"
                           },
                           "description": {
                              "type": "string"
                           },
                           "price": {
                              "type": "number",
                              "format": "float"
                           }
                        }
                     }
                  }
               },
               "404": {
                  "description": "Not found"
               },
               "500": {
                  "description": "Internal server error"
               }
            }
         },
         "post" : {
            "operationId": "services_create",
            "tags": [
               "Services"
            ],
            "description": "Crée un service",
            "parameters": [
               {
                  "name": "service",
                  "in": "body",
                  "required": true,
                  "schema": {
                     "type": "object",
                     "properties": {
                        "service_name": {
                           "type": "string"
                        },
                        "description": {
                           "type": "string"
                        },
                        "price": {
                           "type": "number",
                           "format": "float"
                        }
                     }
                  }
               }
            ],
            "responses": {
               "201": {
                  "description": "successful operation",
                  "schema": {
                     "type": "object",
                     "properties": {
                        "id_service": {
                           "type": "integer"
                        },
                        "service_name": {
                           "type": "string"
                        },
                        "description": {
                           "type": "string"
                        },
                        "price": {
                           "type": "number",
                           "format": "float"
                        }
                     }
                  }
               }
            }
         }
      },
...
   }
}